further improved design
